NeVerimAbime/requirements.txt
Django==3.0.5
asgiref==3.2.7
pip==20.0.2
pytz==2019.3
setuptools==46.1.3
sqlparse==0.3.1
psycopg2==2.8.5
uuid
requests


amqp
anyjson
billiard
celery
django-celery
django-redis
django-redis-cache
kombu
python-memcached
redis
six
vine


django-rest-auth
djangorestframework
django-allauth
djoser
django-cors-headers
djangorestframework_simplejwt
pandas
wordcloud
gunicorn
coverage
django-rest-swagger
drf-yasg

